% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_unquoted_col_name}
\alias{check_unquoted_col_name}
\title{Check that an unquoted column name exists in a data frame}
\usage{
check_unquoted_col_name(x, nm)
}
\arguments{
\item{x}{A data frame.}

\item{nm}{An unquoted column name.}
}
\description{
Check that an unquoted column name exists in a data frame
}
