% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_test_names}
\alias{clean_test_names}
\title{Clean Test Names}
\usage{
clean_test_names(x)
}
\arguments{
\item{x}{A character vector of test names to be cleaned.}
}
\value{
A character vector of cleaned test names.
}
\description{
Cleans test names by removing the standards year (e.g., "(23)") and " CAT" from the string.
}
\examples{
\dontrun{
clean_test_names(c("Gr 4 Reading CAT", "EOC Reading (17)", "Gr 8 Math (16) CAT"))
# Returns: c("Gr 4 Reading", "EOC Reading", "Gr 8 Math")
}
}
