% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{drop_failing_retests}
\alias{drop_failing_retests}
\title{Drop Failing Retests}
\usage{
drop_failing_retests(x)
}
\arguments{
\item{x}{A dataframe, ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}}}
}
\value{
a dataframe with failing retests removed
}
\description{
Filter a student data extract to remove failing retests in alignment with VDOE's steps to calculate annual pass rates. This function assumes that retest status is defined in a column named "retest," where "Y" indicates a retest. It also assumes that performance levels are defined in a column named "performance_level."
}
\examples{
\dontrun{
df_no_fail_retest <- drop_failing_retests(my_data)
}
}
