% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{drop_hs_failing_retests}
\alias{drop_hs_failing_retests}
\title{Drop Failing High School Retests}
\usage{
drop_hs_failing_retests(
  x,
  hs_pattern = "HS$|CAREER",
  retest_col = retest,
  performance_lvl_col = performance_level,
  school_col = school_name
)
}
\arguments{
\item{x}{A dataframe, ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}}.}

\item{hs_pattern}{A character string containing a regular expression used to identify high schools from the school name column. The default is \code{"HS$|CAREER"}.}

\item{retest_col}{The unquoted column name for the retest indicator. Defaults to \code{retest}.}

\item{performance_lvl_col}{The unquoted column name for the performance level. Defaults to \code{performance_level}.}

\item{school_col}{The unquoted column name for the school name. Defaults to \code{school_name}.}
}
\value{
A dataframe with failing high school retests removed.
}
\description{
Filter a student data extract to remove failing retests for high school tests. This is a more specific version of \code{\link[=drop_failing_retests]{drop_failing_retests()}} that targets only high schools, identified by a pattern in the school name.
}
\examples{
\dontrun{
my_data <- ingest_student_data_extract("path/to/my/data.csv")
# Remove failing retests from schools with "HS" or "CAREER" in their name
df_no_hs_fail_retest <- drop_hs_failing_retests(my_data)
}
}
