% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{drop_irw}
\alias{drop_irw}
\title{Drop Int Read Write Tests}
\usage{
drop_irw(x)
}
\arguments{
\item{x}{A dataframe, ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}}}
}
\value{
a dataframe with IRW tests removed
}
\description{
Filter a student data extract to remove integrated reading and writing (IRW) tests. This function assumes that test names are defined in a column named "test_name."
}
\examples{
\dontrun{
df_no_irw <- drop_irw(my_data)
}
}
