% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_exclusions}
\alias{filter_exclusions}
\title{Filter Exclusions}
\usage{
filter_exclusions(x, drop_parent_requested = TRUE, drop_failing_retests = TRUE)
}
\arguments{
\item{x}{A data frame, ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}}, containing student test data.}

\item{drop_parent_requested}{Logical. If \code{TRUE} (the default), exclude tests marked as parent-requested.}

\item{drop_failing_retests}{Logical. If \code{TRUE} (the default), exclude failing retests. See \code{\link[=drop_failing_retests]{drop_failing_retests()}} for details.}
}
\value{
A filtered data frame with the specified exclusions removed.
}
\description{
This function filters a student data extract to exclude certain records based on several criteria,
aligning with VDOE's guidelines for calculating annual pass rates.  Specifically, it removes records where:
\itemize{
\item The grade is "TT" (non-enrolled student).
\item The performance level is not in the set of allowed levels (1, 2, 3, 4, 5, 8).
\item The student is a recently arrived English learner (EL) taking a reading test and has a non-passing performance level (3, 4, 5).
\item The test was parent-requested (if \code{drop_parent_requested} is \code{TRUE}).
\item The test is a failing retest (if \code{drop_failing_retests} is \code{TRUE}).
}
}
