% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_test_performance}
\alias{filter_test_performance}
\title{Title}
\usage{
filter_test_performance(x, type = "best")
}
\arguments{
\item{x}{A dataframe, ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}}}

\item{type}{Character. Either "best" (the default) or "first." "best" will retain only each student's best score on a given test, whereas "first" will retain each student's chronological first score on a given test.}
}
\value{
a dataframe with 1 row per student per test
}
\description{
insert description
}
