% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_writing_extract}
\alias{ingest_writing_extract}
\title{Ingest a Writing Student Data Extract}
\usage{
ingest_writing_extract(path, clean_test_names = TRUE, ...)
}
\arguments{
\item{path}{Character. The path to the file to be ingested.}

\item{clean_test_names}{Logical. If \code{TRUE} (the default), test names will be cleaned using \code{\link[=clean_test_names]{clean_test_names()}}.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A dataframe containing the ingested data.
}
\description{
This is a wrapper around \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}} with \code{writing_extract} set to \code{TRUE}.
It ingests and provides light cleaning for a student data extract for SOL writing tests.
See \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}} for more details on arguments and functionality.
}
\examples{
\dontrun{
p <- "path/to/my/writing_data.csv"
df <- ingest_writing_extract(p)
}
}
