% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_el_status}
\alias{recode_el_status}
\title{Recode English Learner (EL) Status}
\usage{
recode_el_status(el_proficiency_lvl_var)
}
\arguments{
\item{el_proficiency_lvl_var}{A numeric vector representing EL proficiency level. \code{NA} values typically indicate students who are currently classified as EL.}
}
\value{
A character vector with "EL" or "Not_EL" values.
}
\description{
Recodes a variable representing English Learner (EL) proficiency level into text categories indicating EL status.
}
\examples{
\dontrun{
# Assuming you have a data frame called 'student_data' with an 'el_proficiency' column
student_data <- student_data |>
    dplyr::mutate(
        el_status = recode_el_status(el_proficiency)
    )
}
}
