% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_performance_levels}
\alias{recode_performance_levels}
\title{Recode Performance Levels}
\usage{
recode_performance_levels(x)
}
\arguments{
\item{x}{Numeric vector of performance level codes}
}
\value{
A character vector of performance levels
}
\description{
Recode performance levels from numeric codes to text representations
}
\examples{
\dontrun{
# ...ingest and clean data previously
res <- my_sol_data |>
    dplyr::mutate(text_performance_level = recode_performance_levels(performance_level))
}
}
