% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_race_eth}
\alias{recode_race_eth}
\title{Recode Race and Ethnicity Variables}
\usage{
recode_race_eth(eth_var = NULL, race_var = NULL)
}
\arguments{
\item{eth_var}{A character vector representing ethnicity where "Y" indicates Hispanic ethnicity and "N" indicates non-Hispanic ethnicity.}

\item{race_var}{A numeric vector representing race, where the numeric codes correspond to VDOE's race codes.}
}
\value{
A character vector of recoded race/ethnicity values.
}
\description{
Recodes race and ethnicity variables into standardized text categories.
}
\details{
This function uses \code{dplyr::case_when} to recode the input variables.
\itemize{
\item If \code{eth_var} is "Y", the output is "Hispanic".
\item Otherwise, the function checks \code{race_var} for the following codes:
\itemize{
\item 1: "American Indian/Alaska Native"
\item 2: "Asian"
\item 3: "Black"
\item 5: "White"
\item 6: "Hawaiian or Pacific Islander"
\item 7 or greater: "Two or More Races"
}
\item If none of the above conditions are met, the function will return NA.
}
}
\examples{
\dontrun{
# Assuming you have a data frame called 'student_data' with 'ethnicity' and 'race' columns
student_data <- student_data |>
    dplyr::mutate(
        race_ethnicity = recode_race_eth(ethnicity, race)
    )

# Using different column names
student_data <- student_data |>
    dplyr::mutate(
        race_ethnicity = recode_race_eth(hispanic_flag, race_code)
    )
}
}
