% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_sdbq.R
\name{recode_sdbq}
\alias{recode_sdbq}
\title{Recode SDBQ}
\usage{
recode_sdbq(x, collapse_races = FALSE, collapse_thresh = 0.05)
}
\arguments{
\item{x}{data. SDBQ data extract to be passed in.}

\item{collapse_races}{logical. If TRUE, will collapse infrequently occuring races into an "Other" category.}

\item{collapse_thresh}{numeric value between 0 and 1. Races with a rate of occurence less than this will be collapsed into "Other".}
}
\value{
A dataframe with recoded values
}
\description{
This function recodes two columns in the SDBQ data extracts. It recodes \code{response} column to numeric values (1s and 0s), and it recodes the \code{race} column to text rather than codes (e.g. "05" is recoded to "White").

It also provides options to collapse infrequently-occurring races into an "Other" category. You can set the threshold at which races are considered infrequently occuring using the \code{collapse_thresh} argument.
}
\examples{
\dontrun{
library(tidyverse)
#first read in your sdbq extract from Pearson
df <- read_csv("my_sdbq_extract.csv")

#basic version
a <- recode_sdbq(df)

#collapsing race codes with rates less than 10\% into other
b <- recode_sdbq(df, collapse_races = TRUE, collapse_thresh = .1)
}
}
