% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_sped_status}
\alias{recode_sped_status}
\title{Recode SPED Status}
\usage{
recode_sped_status(sped_var)
}
\arguments{
\item{sped_var}{A numeric vector representing SPED status.  \code{NA} or 15 typically indicates "Not SPED".}
}
\value{
A character vector with "Not_SPED" or "SPED" values.
}
\description{
Recodes a variable representing Special Education (SPED) status into text categories.
}
\examples{
\dontrun{
# Assuming you have a data frame called 'student_data' with a 'sped_status' column
student_data <- student_data |>
    dplyr::mutate(
        sped_status_text = recode_sped_status(sped_status)
    )
}
}
