% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{standardize_writing_extract_names}
\alias{standardize_writing_extract_names}
\title{Standardize Writing Extract Names}
\usage{
standardize_writing_extract_names(x)
}
\arguments{
\item{x}{A data frame containing writing test data, typically from an SOL student data extract. Ideally one created by \code{\link[=ingest_student_data_extract]{ingest_student_data_extract()}} where \code{writing_extract} is \code{TRUE}.
It should include the columns \code{performance_level_total_mc_tei_sp} and \code{scaled_score_writing_total}.}
}
\value{
A data frame with standardized column names: \code{performance_level} and \code{test_scaled_score}.
}
\description{
Standardizes some column names in a writing test data frame to align with the structure of non-writing test data frames.
Specifically, it renames \code{performance_level_total_mc_tei_sp} to \code{performance_level} and
\code{scaled_score_writing_total} to \code{test_scaled_score}.  This ensures consistency when
analyzing both writing and non-writing assessment data.
}
\examples{
\dontrun{
writing_data <- ingest_student_data_extract("path/to/data.csv", writing_extract = TRUE)
standardized_data <- standardize_writing_extract_names(writing_data)
}
}
