% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_student_data_extract}
\alias{ingest_student_data_extract}
\title{Ingest Student Data Extract}
\usage{
ingest_student_data_extract(path, ...)
}
\arguments{
\item{path}{Character. The path to the file to be ingested.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
A dataframe containing the ingested data.
}
\description{
Ingest an SOL student data extract (csv file). This should be the exact file downloaded from Pearson
}
